+++ packages/dock-blockchain-modules/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @docknetwork/dock-blockchain-modules
 
+## 0.26.0
+
+### Minor Changes
+
+- Tweaks
+
+### Patch Changes
+
+- Updated dependencies
+  - @docknetwork/credential-sdk@0.38.0
+
 ## 0.25.0
 
 ### Minor Changes